# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Makers Community API",
        "description": "API complète pour la plateforme communautaire Makers Community avec authentification, posts, commentaires, ratings et modération automatique. Cette API permet de gérer une communauté de makers avec toutes les fonctionnalités nécessaires.",
        "contact": {
            "name": "Équipe Makers Community",
            "email": "dev@makerscomm.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement local"
        },
        {
            "url": "https://api.makerscomm.com/api",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/posts/{post}/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Lister les commentaires d'un post",
                "description": "Récupérer tous les commentaires principaux d'un post avec leurs réponses (structure hiérarchique)",
                "operationId": "00a7052dfabc2d7ba47f1f5d5d62f176",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID du post",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Commentaires récupérés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "error": {
                                            "type": "null"
                                        },
                                        "data": {
                                            "properties": {
                                                "post_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "comments": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Comment"
                                                    }
                                                },
                                                "total_comments": {
                                                    "type": "integer",
                                                    "example": 5
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Commentaires récupérés avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Créer un nouveau commentaire",
                "description": "Ajouter un commentaire à un post. Peut être un commentaire principal ou une réponse à un autre commentaire.",
                "operationId": "98a3c321f84f38f3212cd81009382df5",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID du post",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content"
                                ],
                                "properties": {
                                    "content": {
                                        "description": "Contenu du commentaire (max 1000 caractères)",
                                        "type": "string",
                                        "example": "Excellent projet ! Comment avez-vous géré la partie hardware ?"
                                    },
                                    "parent_id": {
                                        "description": "ID du commentaire parent pour une réponse",
                                        "type": "integer",
                                        "example": null,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Commentaire créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "error": {
                                            "type": "null"
                                        },
                                        "data": {
                                            "properties": {
                                                "comment": {
                                                    "$ref": "#/components/schemas/Comment"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Commentaire ajouté avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "title": "Utilisateur",
                "description": "Modèle utilisateur de base",
                "properties": {
                    "id": {
                        "description": "Identifiant unique",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nom complet",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "description": "Adresse email",
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "birthday": {
                        "description": "Date de naissance",
                        "type": "string",
                        "format": "date",
                        "example": "1995-05-15"
                    },
                    "role": {
                        "description": "Rôle utilisateur",
                        "type": "string",
                        "enum": [
                            "user",
                            "admin"
                        ],
                        "example": "user"
                    },
                    "profile_pic": {
                        "description": "Photo de profil",
                        "type": "string",
                        "example": "profile_123.jpg",
                        "nullable": true
                    },
                    "is_active": {
                        "description": "Statut actif du compte",
                        "type": "boolean",
                        "example": true
                    },
                    "email_verified_at": {
                        "description": "Date de vérification email",
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-10-02T14:30:00Z",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-10-01T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-10-02T14:30:00Z"
                    }
                },
                "type": "object"
            },
...
            "Comment": {
                "title": "Commentaire",
                "description": "Commentaire hiérarchique",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "content": {
                        "type": "string",
                        "example": "Excellent travail ! Comment as-tu géré la partie electronique ?"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "post_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "parent_comment_id": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-10-02T15:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-10-02T15:00:00Z"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "replies": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Comment"
                        }
                    }
                },
                "type": "object"
            },
...
            "StandardResponse": {
                "title": "Réponse Standard",
                "description": "Format de réponse standardisé",
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "success",
                            "error"
                        ],
                        "example": "success"
                    },
                    "error": {
                        "example": null,
                        "oneOf": [
                            {
                                "type": "null"
                            },
                            {
                                "type": "string"
                            },
                            {
                                "type": "object"
                            }
                        ]
                    },
                    "data": {
                        "oneOf": [
                            {
                                "type": "null"
                            },
                            {
                                "type": "object"
                            },
                            {
                                "type": "array",
                                "items": {}
                            }
                        ]
                    },
                    "message": {
                        "type": "string",
                        "example": "Opération réussie"
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "title": "Erreur de Validation",
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "error"
                    },
                    "error": {
                        "properties": {
                            "field_name": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "Ce champ est requis."
                                }
                            }
                        },
                        "type": "object"
                    },
                    "data": {
                        "type": "null"
                    },
                    "message": {
                        "type": "string",
                        "example": "Erreur de validation"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token d'authentification Laravel Sanctum. Format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}